feat : sign-up 커밋 (#6)
* fix : UserInfo 엔티티 최종 컬럼으로 수정

* refactor : 기존 코드 주석

* refactor : oauth 폴더 삭제

* feat : 카카오 회원가입 API

* feat : 네이버 회원가입 API

* fix : 카카오 회원가입 API - validation 추가

+++ src/main/java/com/example/ged/utils/JwtService.java
@@ -23,7 +23,7 @@ public class JwtService {
      * @param userIdx
      * @return String
      */
-    public String createJwt(int userIdx) {
+    public String createJwt(Long userIdx) {
         Date now = new Date();
         return Jwts.builder()
                 .claim("userIdx", userIdx)
